Statistics moved to StatisticsController Rename StatisticsDTO to StatisticsInterface
