+++ src/main/kotlin/com/vakzu/musicwars/dto/StatisticsDto.kt
@@ -1,0 +1,8 @@
+package com.vakzu.musicwars.dto
+
+class StatisticsDto(
+    val playedGamesAmount: Int,
+    val winsAmount: Int,
+    val averagePrice: Double,
+    val lastGameTimeStamp: String
+)
